+++ src/parseTree.ml
@@ -55,6 +55,7 @@ and arithmetic_operator =
   | DivEuc
   | Modulo
   | DivMod
+  | ThreeWayCmp
   | ShiftLeft
   | ShiftRight