+++ src/printer_model_scaml.ml
@@ -933,6 +933,14 @@ let pp_model fmt (model : model) =
         in
         pp fmt (l, r)
 
+      | MthreeWayCmp (l, r) ->
+        let pp fmt (l, r) =
+          Format.fprintf fmt "%a <=> %a"
+            f l
+            f r
+        in
+        pp fmt (l, r)
+
       | Mshiftright (l, r) ->
         let pp fmt (l, r) =
           Format.fprintf fmt "%a >> %a"