+++ src/printer_pt.ml
@@ -41,6 +41,7 @@ let e_mult          =  (110, Left)     (* *   *)
 let e_divrat        =  (110, Left)     (* /   *)
 let e_modulo        =  (110, Left)     (* %   *)
 let e_divmod        =  (110, Left)     (* /%  *)
+let e_three_way_cmp =  (112, Left)     (* <=> *)
 let e_left_shift    =  (115, Left)     (* <<  *)
 let e_right_shift   =  (115, Left)     (* >>  *)
 let e_diveuc        =  (115, Left)     (* div *)
@@ -76,6 +77,7 @@ let get_prec_from_operator (op : operator) =
   | Arith DivEuc    -> e_diveuc
   | Arith Modulo    -> e_modulo
   | Arith DivMod    -> e_divmod
+  | Arith ThreeWayCmp -> e_three_way_cmp
   | Arith ShiftLeft -> e_left_shift
   | Arith ShiftRight-> e_right_shift
   | Unary Uminus    -> e_minus
@@ -206,8 +208,9 @@ let arithmetic_operator_to_str op =
   | DivEuc -> "div"
   | Modulo -> "%"
   | DivMod -> "/%"
-  | ShiftLeft  -> "<-<"
-  | ShiftRight -> ">->"
+  | ThreeWayCmp  -> "<=>"
+  | ShiftLeft  -> "<<|"
+  | ShiftRight -> "|>>"
 
 let unary_operator_to_str op =
   match op with