+++ CHANGES.md
@@ -8,8 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
  - records shaping with `as`
  - add `parameter`
- - add left and right shift operators (`<-<` and `>->`)
- - add divmod operator (EDIV in michelson)
+ - add left and right shift operators (`<<|` and `|>>`)
+ - add divmod and three-way comparison operators (resp. EDIV and COMPARE in michelson)
 
 ### Changed
  - handle decimal for percent and currency literals